improve mesh between small spacings
Previously, when a large distance was surrounding by fine features,
the distance used the spacings from the fine features. This led to
unnecessarily fine meshes in this distances and long simulations. The
mesh is now able to expand and contract again within these distances
to reduce the number of mesh lines.