mesh.py: pml smoothing now works, but sometimes clips structures
There shouldn't be any major issues with clipping structures inside a
PML. However, it's visually a bit unsettling for users to see the mesh
not encompass their entire structure. Therefore I'm going to modify
this to keep the full structure and still smooth the PML mesh. I'm
keeping this commit in case at some future point it becomes clear that
this implementation was better.